LayoutTests/platform/ios-simulator/editing/secure-input/password-input-focusing-to-different-frame-expected.txt

Verify that changing focus to a different frame updates secure input state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial state, no selection:
PASS testRunner.secureEventInputIsEnabled is false

A password input is focused:
FAIL document.activeElement should be [object HTMLInputElement]. Was [object HTMLBodyElement].
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.

A regular text input is focused with a mouse click:
FAIL frames[0].document.activeElement should be [object HTMLInputElement]. Was [object HTMLBodyElement].
PASS testRunner.secureEventInputIsEnabled is false

A password input is focused again with a mouse click:
FAIL document.activeElement should be [object HTMLInputElement]. Was [object HTMLBodyElement].
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.

A regular text input is focused with a tab:
FAIL frames[0].document.activeElement should be [object HTMLInputElement]. Was [object HTMLBodyElement].
PASS testRunner.secureEventInputIsEnabled is false

A password input is focused again with a tab:
FAIL document.activeElement should be [object HTMLInputElement]. Was [object HTMLBodyElement].
FAIL testRunner.secureEventInputIsEnabled should be true. Was false.

Focus is made empty:
PASS document.activeElement is document.body
PASS testRunner.secureEventInputIsEnabled is false
PASS successfullyParsed is true

TEST COMPLETE
 
